chore(cmd/trait): refactor cmd & trait packages
