+++ pkg/trait/util_test.go
@@ -21,86 +21,7 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-
-	v1 "github.com/apache/camel-k/pkg/apis/camel/v1"
 )
-
-func TestCollectConfigurationValues(t *testing.T) {
-	e := Environment{
-		Integration: &v1.Integration{
-			Spec: v1.IntegrationSpec{
-				Configuration: []v1.ConfigurationSpec{
-					{Type: "configmap", Value: "my-cm-integration"},
-					{Type: "env", Value: "my-env-integration"},
-				},
-			},
-		},
-		IntegrationKit: &v1.IntegrationKit{
-			Spec: v1.IntegrationKitSpec{
-				Configuration: []v1.ConfigurationSpec{
-					{Type: "configmap", Value: "my-cm-kit"},
-					{Type: "property", Value: "my-p-kit"},
-				},
-			},
-		},
-		Platform: &v1.IntegrationPlatform{
-			Spec: v1.IntegrationPlatformSpec{
-				Configuration: []v1.ConfigurationSpec{
-					{Type: "configmap", Value: "my-cm-platform"},
-					{Type: "secret", Value: "my-secret-platform"},
-					{Type: "property", Value: "my-p-platform"},
-					{Type: "env", Value: "my-env-platform"},
-				},
-			},
-		},
-	}
-	e.Platform.ResyncStatusFullConfig()
-
-	assert.Contains(t, e.collectConfigurationValues("configmap"), "my-cm-integration")
-	assert.Contains(t, e.collectConfigurationValues("secret"), "my-secret-platform")
-	assert.Contains(t, e.collectConfigurationValues("property"), "my-p-kit")
-	assert.Contains(t, e.collectConfigurationValues("env"), "my-env-integration")
-}
-
-func TestCollectConfigurationPairs(t *testing.T) {
-	e := Environment{
-		Integration: &v1.Integration{
-			Spec: v1.IntegrationSpec{
-				Configuration: []v1.ConfigurationSpec{
-					{Type: "property", Value: "p1=integration"},
-					{Type: "property", Value: "p4=integration"},
-				},
-			},
-		},
-		IntegrationKit: &v1.IntegrationKit{
-			Spec: v1.IntegrationKitSpec{
-				Configuration: []v1.ConfigurationSpec{
-					{Type: "property", Value: "p1=kit"},
-					{Type: "property", Value: "p2=kit"},
-				},
-			},
-		},
-		Platform: &v1.IntegrationPlatform{
-			Spec: v1.IntegrationPlatformSpec{
-				Configuration: []v1.ConfigurationSpec{
-					{Type: "property", Value: "p1=platform"},
-					{Type: "property", Value: "p2=platform"},
-					{Type: "property", Value: "p3=platform"},
-					{Type: "property", Value: "p4=platform"},
-				},
-			},
-		},
-	}
-	e.Platform.ResyncStatusFullConfig()
-
-	pairs := e.collectConfigurationPairs("property")
-	assert.Equal(t, pairs, []variable{
-		{Name: "p1", Value: "integration"},
-		{Name: "p2", Value: "kit"},
-		{Name: "p3", Value: "platform"},
-		{Name: "p4", Value: "integration"},
-	})
-}
 
 func TestBoolPointerFunctions(t *testing.T) {
 	trueP := BoolP(true)
